requirements.txt
langchain-core>=0.2.28,<0.3.0
langchain-google-genai>=1.0.1,<2.0.0
langchain-community>=0.1.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
pydantic>=1.10,<2.0
